requirements.txt
gym==0.17.2
requests==2.22.0
numpy==1.18.5
tensorflow==2.3.0
python-dotenv==0.14.0
